refactor: move Keypair class to top level
