+++ tests/fixtures.py
@@ -1,6 +1,6 @@
 # flake8: noqa: E501
 # pylint: disable=line-too-long,missing-module-docstring
-from kinetic_sdk import Keypair
+from kinetic_sdk.keypair import Keypair
 
 # fmt: off
 ALICE_KEY = [205, 213, 7, 246, 167, 206, 37, 209, 161, 129, 168, 160, 90, 103, 198, 142, 83, 177, 214, 203, 80, 29, 71, 245, 56, 152, 15, 8, 235, 174, 62, 79, 138, 198, 145, 111, 119, 33, 15, 237, 89, 201, 122, 89, 48, 221, 224, 71, 81, 128, 45, 97, 191, 105, 37, 228, 243, 238, 130, 151, 53, 221, 172, 125]
@@ -17,11 +17,9 @@ CHARLIE_ACCOUNT = "CharYfTvJSiH6LtDpkGUiVVZmeCn5Cenu2TzdJSbDJnG"
 DEFAULT_MINT = "MoGaMuJnB3k8zXjBYBnHxHG47vWcW3nyb7bFYvdVzek"
 FEE_PAYER = "oWNEYV3aMze3CppdgyFAiEj9xUJXkn85es1KscRHt8m"
 
-TEST_MNEMONIC_12 = "field pool drill reward habit engine useless mind hybrid tiny lamp key"
+TEST_MNEMONIC_12 = "pill tomorrow foster begin walnut borrow virtual kick shift mutual shoe scatter"
 TEST_MNEMONIC_12_PUBLIC_KEY = "5F86TNSTre3CYwZd1wELsGQGhqG2HkN3d8zxhbyBSnzm"
 
-
-TEST_MNEMONIC_12 = "pill tomorrow foster begin walnut borrow virtual kick shift mutual shoe scatter"
 TEST_MNEMONIC_12_KEYPAIR = {
     "mnemonic": TEST_MNEMONIC_12,
     "public_key": TEST_MNEMONIC_12_PUBLIC_KEY,
